Move probe_records to usdt
