+++ dusty/Cargo.toml
@@ -10,6 +10,5 @@ rust-version = "1.70.0"
 [dependencies]
 clap = { version = "4.4.11", features = ["derive"] }
 dof = { path = "../dof", features = ["des"] }
-goblin = { version = "0.6", features = ["elf32", "elf64"] }
-memmap = "0.7"
+usdt = { path = "../usdt", features = ["asm"] }
 usdt-impl = { path = "../usdt-impl", features = ["des"] }
