+++ dusty/src/main.rs
@@ -15,14 +15,8 @@
 // limitations under the License.
 
 use clap::Parser;
-use dof::Section;
-use goblin::Object;
-use memmap::Mmap;
-use memmap::MmapOptions;
-use std::fs::File;
-use std::fs::OpenOptions;
-use std::path::Path;
 use std::path::PathBuf;
+use usdt::probe_records;
 use usdt_impl::Error as UsdtError;
 
 /// Inspect data related to USDT probes in object files.
@@ -82,82 +76,5 @@ fn main() {
         Err(e) => {
             println!("Failed to parse probe information, {:?}", e);
         }
-    }
-}
-
-// Extract probe records from the given file, if possible.
-pub(crate) fn probe_records<P: AsRef<Path>>(path: P) -> Result<Section, UsdtError> {
-    let file = OpenOptions::new().read(true).create(false).open(path)?;
-    let (offset, len) = locate_probe_section(&file).ok_or(UsdtError::InvalidFile)?;
-
-    // Remap only the probe section itself as mutable, using a private
-    // copy-on-write mapping to avoid writing to disk in any circumstance.
-    let mut map = unsafe { MmapOptions::new().offset(offset).len(len).map_copy(&file)? };
-    usdt_impl::record::process_section(&mut map, /* register = */ false)
-}
-
-// Return the offset and size of the file's probe record section, if it exists.
-fn locate_probe_section(file: &File) -> Option<(u64, usize)> {
-    let map = unsafe { Mmap::map(file) }.ok()?;
-    match Object::parse(&map).ok()? {
-        Object::Elf(object) => {
-            // Try to find our special `set_dtrace_probes` section from the section headers. These
-            // may not exist, e.g., if the file has been stripped. In that case, we look for the
-            // special __start and __stop symbols themselves.
-            if let Some(section) = object.section_headers.iter().find(|header| {
-                if let Some(name) = object.shdr_strtab.get_at(header.sh_name) {
-                    name == "set_dtrace_probes"
-                } else {
-                    false
-                }
-            }) {
-                Some((section.sh_offset, section.sh_size as usize))
-            } else {
-                // Failed to look up the section directly, iterate over the symbols.
-                let mut bounds = object.syms.iter().filter(|symbol| {
-                    matches!(
-                        object.strtab.get_at(symbol.st_name),
-                        Some("__start_set_dtrace_probes") | Some("__stop_set_dtrace_probes")
-                    )
-                });
-
-                if let (Some(start), Some(stop)) = (bounds.next(), bounds.next()) {
-                    Some((start.st_value, (stop.st_value - start.st_value) as usize))
-                } else {
-                    None
-                }
-            }
-        }
-        Object::Mach(goblin::mach::Mach::Binary(object)) => {
-            // Try to find our special `__dtrace_probes` section from the section headers.
-            for (section, _) in object.segments.sections().flatten().flatten() {
-                if section.sectname.starts_with(b"__dtrace_probes") {
-                    return Some((section.offset as u64, section.size as usize));
-                }
-            }
-
-            // Failed to look up the section directly, iterate over the symbols.
-            if let Some(syms) = object.symbols {
-                let mut bounds = syms.iter().filter_map(|symbol| {
-                    if let Ok((name, nlist)) = symbol {
-                        if name.contains("__dtrace_probes") {
-                            Some(nlist.n_value)
-                        } else {
-                            None
-                        }
-                    } else {
-                        None
-                    }
-                });
-                if let (Some(start), Some(stop)) = (bounds.next(), bounds.next()) {
-                    Some((start, (stop - start) as usize))
-                } else {
-                    None
-                }
-            } else {
-                None
-            }
-        }
-        _ => None,
     }
 }
